agentic_test_gen/poc_requirements.txt
# PoC Validation Requirements
# Additional packages needed for Proof of Concept validation

# LLM Integration
openai>=1.0.0

# Environment Variables
python-dotenv>=1.0.0

# Natural Language Processing for Quality Validation
nltk>=3.8
scikit-learn>=1.3.0

# Async Processing
asyncio

# Data Processing
numpy>=1.24.0

# Standard library dependencies (should be available)
# json, re, time, statistics, pathlib, dataclasses